requirements.txt
absl-py==2.0.0
appnope==0.1.3
asttokens==2.4.1
backcall==0.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
google-auth==2.24.0
google-auth-oauthlib==1.0.0
grpcio==1.59.3
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
zipp==3.17.0
